ci: publish docs on any change (#26)
Remove filter that publishes docs on changes to `mkdocs.yml` and `docs/**` only. Required for docs to be published on changes to `README.md` or `CHANGELOG.md` - both files are rendered on our docs site.

Instead of updating the filter, remove it so that we don't have to maintain it whenever our MkDocs site references files outside of the `docs` directory.